fix(lint): ensure golangci-lint works with go1.23 and fix lint errors

+++ internal/resources/cloud/resource_cloud_stack_service_account_token.go
@@ -94,7 +94,7 @@ func stackServiceAccountTokenCreate(ctx context.Context, d *schema.ResourceData,
 
 	req := gcom.PostInstanceServiceAccountTokensRequest{
 		Name:          d.Get("name").(string),
-		SecondsToLive: common.Ref(int32(d.Get("seconds_to_live").(int))),
+		SecondsToLive: common.Ref(d.Get("seconds_to_live").(int32)),
 	}
 
 	resp, _, err := cloudClient.InstancesAPI.PostInstanceServiceAccountTokens(ctx, stackSlug, strconv.FormatInt(serviceAccountID, 10)).
